requirements.txt
Flask==2.0.0
gunicorn==20.1.0
SQLAlchemy==1.4.15
pg8000==1.19.5
google-cloud-secret-manager==2.4.0
